Release build 129 - fixed types.

+++ ScaryFlight/ScaryFlight/Classes/Scenes/EABaseGameScene.h
@@ -13,6 +13,6 @@
 
 @property (nonatomic, strong) EAHero *hero;
 
-- (void)addBottomPipe:(float)centerY;
+- (void)addBottomPipe:(CGFloat)centerY;
 
 @end
