+++ ScaryFlight/ScaryFlight/Classes/Sprites/EAHero.m
@@ -17,10 +17,10 @@ static CGFloat const kHeroDirection = 28.5f;
     NSParameterAssert(yLimit > 0.0f);
     
     if (self.position.y < yLimit - self.size.height / 2.0f) { // <-- avoid hero to fly away from top of screen
-        float heroDirection = self.zRotation + (float)M_PI_2;
+        CGFloat heroDirection = self.zRotation + (CGFloat)M_PI_2;
         self.physicsBody.velocity = CGVectorMake(0.0f, 0.0f);
-        [self.physicsBody applyImpulse:CGVectorMake(kHeroDirection * cosf(heroDirection),
-                                                    kHeroDirection * sinf(heroDirection))];
+        [self.physicsBody applyImpulse:CGVectorMake(kHeroDirection * cosf((float)heroDirection),
+                                                    kHeroDirection * sinf((float)heroDirection))];
     }
     
     [self runAction:[SKAction playSoundFileNamed:@"Jump.wav" waitForCompletion:YES]];
